fix(conflicts): fix merge conflicts and minor bugs
