+++ src/components/App/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import HomePage from '../../pages/HomePage/HomePage';
-import AboutPage from '../../pages/AboutPage/AboutPage';
+import CarWashPage from '../../pages/CarWashPage/CarWashPage';
 
 import Footer from '../Sections/Footer/Footer';
 import Header from '../Sections/Header/Header';
@@ -12,7 +12,7 @@ function App() {
 			<main id="main" style={{ backgroundColor: '#ECECEC' }}>
 				<Routes>
 					<Route path="/" element={<HomePage />} />
-					<Route path="/carwashes/:id" element={<AboutPage />} />
+					<Route path="/carwashes/:id" element={<CarWashPage />} />
 					<Route path="*" element={<Navigate to="/" replace />} />
 				</Routes>
 			</main>
