+++ src/pages/CarWashPage/CarWashPage.module.css
